+++ src/codemagic/tools/_app_store_connect/action_groups/__init__.py
@@ -1,4 +1,5 @@
 from .apps_action_group import AppsActionGroup
 from .beta_app_review_submissions_action_group import BetaAppReviewSubmissionsActionGroup
 from .beta_build_localizations_action_group import BetaBuildLocalizationsActionGroup
+from .beta_groups_action_group import BetaGroupsActionGroup
 from .builds_action_group import BuildsActionGroup