+++ tests/tools/app_store_connect/test_publish_action.py
@@ -181,3 +181,35 @@ def test_no_skip_package_validation_argument_from_env(cli_argument_group):
         PublishArgument.SKIP_PACKAGE_VALIDATION.from_args(args)
     print(error_info.value)
     assert str(error_info.value) == 'Provided value "False" is not valid'
+
+
+def test_add_build_to_beta_groups(publishing_namespace_kwargs):
+    cli_args = argparse.Namespace(**publishing_namespace_kwargs)
+    with mock.patch.object(AppStoreConnect, 'find_paths') as mock_find_paths, \
+            mock.patch.object(AppStoreConnect, '_get_publishing_application_packages') as mock_get_packages, \
+            mock.patch.object(AppStoreConnect, '_upload_artifact_with_altool') as mock_upload, \
+            mock.patch.object(AppStoreConnect, '_validate_artifact_with_altool') as mock_validate, \
+            mock.patch.object(AppStoreConnect, '_get_uploaded_build_application') as mock_get_app, \
+            mock.patch.object(AppStoreConnect, '_get_uploaded_build') as mock_get_build, \
+            mock.patch.object(AppStoreConnect, 'submit_to_testflight') as mock_submit_to_testflight, \
+            mock.patch.object(AppStoreConnect, 'add_build_to_beta_groups') as mock_add_build_to_beta_groups:
+        ipa_path = pathlib.Path('app.ipa')
+        mock_find_paths.return_value = [ipa_path]
+        mock_get_packages.return_value = [mock.create_autospec(Ipa, instance=True, path=ipa_path)]
+        build = mock.Mock(id='1525e3c9-3015-407a-9ba5-9addd2558224')
+        mock_get_app.return_value = mock.Mock(id='1525e3c9-3015-407a-9ba5-9addd2558224')
+        mock_get_build.return_value = build
+
+        beta_group_names = ['Test group 1', 'Test group 2']
+        patterns = [pathlib.Path('path.pattern')]
+        AppStoreConnect.from_cli_args(cli_args).publish(
+            application_package_path_patterns=patterns,
+            submit_to_testflight=True,
+            beta_group_names=beta_group_names,
+        )
+
+        mock_get_packages.assert_called_with(patterns)
+        mock_validate.assert_called()
+        mock_upload.assert_called()
+        mock_submit_to_testflight.assert_called_with(build.id, None)
+        mock_add_build_to_beta_groups.assert_called_with(build.id, beta_group_names)