+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+Version 0.10.0
+-------------
+
+**Features**
+
+- New `app-store-connect beta-groups` action set to add and remove TestFlight builds from groups of beta testers.
+- `app-store-connect publish` action accepts multiple Beta group names under `--beta-group` key. The uploaded TestFlight build will be made available to the specified groups of beta testers.
+
+**Development / Docs**
+
+- Update `app-store-connect publish` action docs.
+
 Version 0.9.8
 -------------